pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "sogon"
description = "AI-powered subtitle generator from YouTube URLs or local media files"
authors = [
  { name="Changyong Um", email="eum6211@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License"
]
dependencies = [
    "groq>=0.26.0",
    "openai>=1.58.1",
    "python-dotenv>=1.0.0",
    "yt-dlp>=2024.3.10",
    "pydub>=0.25.1",
    "tqdm>=4.67.1",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "pydantic-settings>=2.9.1",
    "typer>=0.12.0",
    "psutil>=7.1.1",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]

dynamic = ["version"]

[tool.hatch.version]
source = "vcs"

[tool.hatch.version.vcs]
tag-pattern = "v(?P<version>.+)" 

[project.scripts]
sogon = "sogon.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["sogon"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
]
local = [
    "stable-ts>=2.17.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "huggingface-hub>=0.19.0",
    "psutil>=5.9.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.13",
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "coverage>=7.9.1",
]

[tool.uv]
package = true

[project.urls]
Repository = "https://github.com/e7217/sogon"
